pyproject.toml
[tool.poetry]
name = "poetry add fastapi sqlalchemy alembic psycopg2 pydantic pytest ruff taskipy"
version = "0.1.0"
description = "app posto"
authors = ["crisleigomessilva <crislei@sousys.com.br>"]
license = "0.0.1"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
psycopg = "^3.2.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
requests = "^2.32.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pdfplumber = "^0.11.4"
python-multipart = "^0.0.20"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.8.4"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "uvicorn app.main:app --reload"
lint = "ruff check ."
test = "pytest tests/"
